requirements.txt
timeloop
python-telegram-bot
pandas
beautifulsoup4
requests
PTable
matplotlib
geopandas
geopy
descartes
pillow